+++ .gitignore
@@ -1,5 +1,2 @@
 node_modules
-msvs/x64
-msvs/Release
-msvs/.vs
-bin
+build/
