+++ lib/helpers.js
