+++ lib/index.js
@@ -47,32 +47,10 @@ const platformFunctions = {
       return '';
     }
   },
-  win32: (() => {
-    const readCwd = helpers.bin('readCwd.exe');
-    const readCwd32 = helpers.bin('readCwd32.exe');
-    let isPlatform64Bit = null;
-    const getIsPlatform64Bit = async () => {
-      const { stdout } = await helpers.exec(
-        '[Environment]::Is64BitOperatingSystem',
-        { shell: 'powershell.exe' }
-      );
-
-      return stdout.startsWith('True');
-    };
-
-    return async (processId) => {
-      if (isPlatform64Bit === null) {
-        isPlatform64Bit = await getIsPlatform64Bit();
-      }
-
-      try {
-        const { stdout } = await helpers.execFile(isPlatform64Bit ? readCwd : readCwd32, [processId]);
-        return stdout.trim();
-      } catch (error) {
-        return '';
-      }
-    };
-  })()
+  win32: (processId) => {
+    let addon = require('../build/Release/readCwd.node') ?? require('../build/Debug/readCwd.node');
+    return addon.readCwd(processId);
+  }
 };
 
 module.exports = platformFunctions[process.platform];
