--- msvs/readCwd.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.1500
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "readCwd", "readCwd.vcxproj", "{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Debug|x64.ActiveCfg = Debug|x64
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Debug|x64.Build.0 = Debug|x64
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Debug|x86.ActiveCfg = Debug|Win32
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Debug|x86.Build.0 = Debug|Win32
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Release|x64.ActiveCfg = Release|x64
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Release|x64.Build.0 = Release|x64
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Release|x86.ActiveCfg = Release|Win32
-		{47704AD2-D7C5-40DA-8A39-6F6A2FF88C4B}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7211B7ED-3EDE-4B59-9B70-F6762054ACB9}
-	EndGlobalSection
-EndGlobal
